Patch to handle regression caused when no address is returned
Getting address from domiffaddress and if adress is not found throws exception VMIPAddressMissingError in both ipv4 and ipv6 case

If the ip_version is ipv6, then it raises exception VMIPAddressMissingError, and going to retry logic to try getting the ip address.
Even for ipv4 , if address is not found then it raises exception VMIPAddressMissingError and then retries will happen.

Signed-off-by: Tasmiya Nalatwad <Tasmiya.Nalatwad@ibm.com>