Add wait time for mk_label check
When the script tries to label a newly created partition,
it might fail saying the partition is not available, as there
is a slight delay between defining the partition and it being
available in the block devices list. This PR helps to fix that
by adding a wait statement with a 10 second timeout.

Signed-off-by: Aniket Sahu <asahu1x@linux.ibm.com>